Re introduced the flickr search which is not useable in the ui. Created class WebSource to use flickrSearch. WebSource downloads with 4 threads. Renamed frontEnd.py to control.py as it is not really a front end as much as it is a hub.
Created method "appendSearch" in control which pulls up and plots a set of images and allows the user to plot circles on a map next to this set.
Circles can continue to be added until the user wants to begin a search of all images within the radius specified by the plotted circles.
This is plotted in an html file 'temp.html'.
This uses many methods from multiple files.

Added horizon detection to the display.

note: This might be buggy as control.py is running things while also being imported by display. I am committing for version control purposes.

Todo: Comment code and clean for others to read and use utility functions.
Todo: Create modular front end linked to display